CMake based configuration and building now fully implemented. Makefiles have been removed as they are now obsolete.
Addition of the -pedantic flag to the c++ compilation necessitated some
chagnes to how some variables are defined. This has also made clear some
other issues that were fixed.

Code::Blocks will no longer be able to make the program, but this is
an issue for a small subset of users.


Former-commit-id: eed584dc15553127b4bbdb84b27990f79d827f26

+++ src/core/PixieStd.cpp
@@ -491,7 +491,7 @@ extern "C" void hissub_(unsigned short *ibuf[],unsigned short *nhw)
         /* while the current location in the buffer has not gone beyond the end
          * of the buffer (ignoring the last three delimiters,
          * continue reading */
-        while (nWords < (nhw[0]/2 - 6)) {
+        while (nWords < (unsigned int)(nhw[0]/2 - 6)) {
             /*
             Retrieve the record length and the vsn number
             */
@@ -590,7 +590,7 @@ extern "C" void hissub_(unsigned short *ibuf[],unsigned short *nhw)
                 break;
             }
         } // while still have words
-        if (nWords > nhw[0] / 2 - 6) {
+        if (nWords > (unsigned int)nhw[0] / 2 - 6) {
             ss << "This actually happens!";
             messenger.run_message(ss.str());
             ss.str("");
